Vim: swap p and P in visual mode
Often times I want to replace a whole line by another multiple times,
but regular p in visual mode will change the content of the unnamed
register to the previously selected text, and prevent repeating the
change, and breaks the muscle memory of pasting in normal mode with p
without impacting the unnamed register.

P is also mapped on the occasions I do want to swap the unnamed register
with the selection.

See also :help put-Visual-mode